Add level resource data and readme for documentation
We talked earlier this morning about the scope for the levels for this prototype. We are going to do 5 levels that all stack on top of eachother and the player can fall through and travel up the levels. Progression will involve needing to fall through certain holes to find ways to specific keys.